+++ src/oihana/core/arrays/stub.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace oihana\core\arrays ;
+
+/**
+ * This function return an empty array.
+ * @return array The new empty array.
+ */
+function stub() :array
+{
+    return [];
+}